+++ spec/fixtures/models/person.rb
@@ -1,7 +1,7 @@
 require 'cat'
 
-class Person < Hash
-  include ::CouchRest::Model::CastedModel
+class Person
+  include ::CouchRest::Model::Embeddable
   property :pet, Cat
   property :name, [String]